Merge NuttX 11.0.0

+++ arch/risc-v/src/rv32m1/rv32m1_start.c
@@ -29,7 +29,7 @@
 #include <nuttx/init.h>
 #include <arch/board/board.h>
 
-#include "riscv_arch.h"
+#include "riscv_internal.h"
 #include "rv32m1_clockconfig.h"
 #include "rv32m1.h"
 #include "rv32m1_gpio.h"
@@ -60,7 +60,7 @@
  * address.
  */
 
-uint32_t g_idle_topstack = RV32M1_IDLESTACK_TOP;
+uintptr_t g_idle_topstack = RV32M1_IDLESTACK_TOP;
 
 /****************************************************************************
  * Public Functions
